refactor(smart-wallet): move game1 proposal fixtures from test to tools
