+++ packages/builders/scripts/smart-wallet/build-game1-start.js
@@ -6,7 +6,7 @@
  */
 
 import { makeHelpers } from '@agoric/deploy-script-support';
-import { getManifestForGame1 } from '@agoric/smart-wallet/test/start-game1-proposal.js';
+import { getManifestForGame1 } from '@agoric/smart-wallet/tools/fixtures/start-game1-proposal.js';
 import { buildBundlePath } from '../lib/build-bundle.js';
 
 /**
@@ -18,18 +18,22 @@ import { buildBundlePath } from '../lib/build-bundle.js';
 const game1ProposalBuilder = async ({ publishRef, install }) => {
   const game1Path = await buildBundlePath(
     import.meta.url,
-    '@agoric/smart-wallet/test/gameAssetContract.js',
+    '@agoric/smart-wallet/tools/fixtures/gameAssetContract.js',
     'game1',
   );
   return harden({
-    sourceSpec: '@agoric/smart-wallet/test/start-game1-proposal.js',
+    sourceSpec: '@agoric/smart-wallet/tools/fixtures/start-game1-proposal.js',
     getManifestCall: [
       getManifestForGame1.name,
       {
         game1Ref: publishRef(
-          install('@agoric/smart-wallet/test/gameAssetContract.js', game1Path, {
-            persist: true,
-          }),
+          install(
+            '@agoric/smart-wallet/tools/fixtures/gameAssetContract.js',
+            game1Path,
+            {
+              persist: true,
+            },
+          ),
         ),
       },
     ],
