+++ packages/deploy-script-support/README.md
@@ -21,18 +21,18 @@ Here's a simple example:
 
 ```js
 import { makeHelpers } from '@agoric/deploy-script-support';
-import { getManifestForGame1 } from '@agoric/smart-wallet/test/start-game1-proposal.js';
+import { getManifestForGame1 } from '@agoric/smart-wallet/tools/fixtures/start-game1-proposal.js';
 
 /** @type {import('@agoric/deploy-script-support/src/externalTypes.js').CoreEvalBuilder} */
 const game1ProposalBuilder = async ({ publishRef, install }) => {
   return harden({
-    sourceSpec: '@agoric/smart-wallet/test/start-game1-proposal.js',
+    sourceSpec: '@agoric/smart-wallet/tools/fixtures/start-game1-proposal.js',
     getManifestCall: [
       getManifestForGame1.name,
       {
         game1Ref: publishRef(
           install(
-            '@agoric/smart-wallet/test/gameAssetContract.js',
+            '@agoric/smart-wallet/tools/fixtures/gameAssetContract.js',
             '../bundles/bundle-game1.js',
             { persist: true },
           ),
@@ -49,7 +49,7 @@ export default async (homeP, endowments) => {
 ```
 
 This CoreEvalBuilder returns an object whose "sourceSpec" indicates that the
-proposal to run is "@agoric/smart-wallet/test/start-game1-proposal.js" and whose
+proposal to run is "@agoric/smart-wallet/tools/fixtures/start-game1-proposal.js" and whose
 "getManifestCall" is a [functionName, ...args] array describing an invocation of
 `getManifestForGame1` exported from that file which is expected to return an
 object including a "manifest" as described below (but the actual invocation will
