+++ packages/smart-wallet/package.json
@@ -45,6 +45,7 @@
   },
   "files": [
     "src/",
+    "tools/",
     "source-spec-registry.*"
   ],
   "keywords": [],
