+++ packages/smart-wallet/test/invoke-result.test.js
@@ -6,7 +6,7 @@ import { objectMap } from '@agoric/internal';
 import { makeCopyBag } from '@agoric/store';
 import { E, passStyleOf } from '@endo/far';
 import { deploy } from '../tools/wf-tools.js';
-import * as gameExports from './gameAssetContract.js';
+import * as gameExports from '../tools/fixtures/gameAssetContract.js';
 import * as priceExports from './wallet-fun.contract.js';
 
 /**
