+++ packages/smart-wallet/tools/fixtures/start-game1-proposal.js
@@ -1,0 +1,122 @@
+// @ts-check
+import { Fail } from '@endo/errors';
+import { E } from '@endo/far';
+import { makeMarshal } from '@endo/marshal';
+import { AmountMath } from '@agoric/ertp/src/amountMath.js';
+
+/**
+ * @import {BootstrapManifest} from '@agoric/vats/src/core/lib-boot.js';
+ * @import {StorageNode} from '@agoric/internal/src/lib-chainStorage.js';
+ * @import {ERef} from '@agoric/vow';
+ * @import {BootstrapPowers} from '@agoric/vats/src/core/types.js';
+ */
+
+// vstorage paths under published.*
+const BOARD_AUX = 'boardAux';
+
+const marshalData = makeMarshal(_val => Fail`data only`);
+
+const IST_UNIT = 1_000_000n;
+const CENT = IST_UNIT / 100n;
+
+/**
+ * Make a storage node for auxiliary data for a value on the board.
+ *
+ * @param {ERef<StorageNode>} chainStorage
+ * @param {string} boardId
+ */
+const makeBoardAuxNode = async (chainStorage, boardId) => {
+  const boardAux = E(chainStorage).makeChildNode(BOARD_AUX);
+  return E(boardAux).makeChildNode(boardId);
+};
+
+const publishBrandInfo = async (chainStorage, board, brand) => {
+  const [id, displayInfo] = await Promise.all([
+    E(board).getId(brand),
+    E(brand).getDisplayInfo(),
+  ]);
+  const node = makeBoardAuxNode(chainStorage, id);
+  const aux = marshalData.toCapData(harden({ displayInfo }));
+  await E(node).setValue(JSON.stringify(aux));
+};
+
+/**
+ * Core eval script to start contract
+ *
+ * @param {BootstrapPowers} permittedPowers
+ */
+export const startGameContract = async permittedPowers => {
+  const {
+    consume: { agoricNames, board, chainStorage, startUpgradable, zoe },
+    brand: {
+      // @ts-expect-error dynamic extension to promise space
+      produce: { Place: producePlaceBrand },
+    },
+    issuer: {
+      // @ts-expect-error dynamic extension to promise space
+      produce: { Place: producePlaceIssuer },
+    },
+    instance: {
+      // @ts-expect-error dynamic extension to promise space
+      produce: { game1: produceInstance },
+    },
+  } = permittedPowers;
+
+  const istBrand = await E(agoricNames).lookup('brand', 'IST');
+  const ist = {
+    brand: istBrand,
+  };
+  // NOTE: joinPrice could be configurable
+  const terms = { joinPrice: AmountMath.make(ist.brand, 25n * CENT) };
+
+  // agoricNames can be updated on each bootstrap run; the promise-space
+  // producers resolve only once and then remain fixed.
+  const installation = await E(agoricNames).lookup('installation', 'game1');
+
+  const { instance } = await E(startUpgradable)({
+    installation,
+    label: 'game1',
+    terms,
+  });
+  const {
+    brands: { Place: brand },
+    issuers: { Place: issuer },
+  } = await E(zoe).getTerms(instance);
+
+  produceInstance.reset();
+  produceInstance.resolve(instance);
+
+  producePlaceBrand.reset();
+  producePlaceIssuer.reset();
+  producePlaceBrand.resolve(brand);
+  producePlaceIssuer.resolve(issuer);
+
+  await publishBrandInfo(chainStorage, board, brand);
+};
+
+/** @type {BootstrapManifest} */
+const gameManifest = {
+  [startGameContract.name]: {
+    consume: {
+      agoricNames: true,
+      board: true, // to publish boardAux info for game NFT
+      chainStorage: true, // to publish boardAux info for game NFT
+      startUpgradable: true, // to start contract and save adminFacet
+      zoe: true, // to get contract terms, including issuer/brand
+    },
+    installation: { consume: { game1: true } },
+    issuer: { produce: { Place: true } },
+    brand: { produce: { Place: true } },
+    instance: { produce: { game1: true } },
+  },
+};
+harden(gameManifest);
+
+export const getManifestForGame1 = ({ restoreRef }, { game1Ref }) => {
+  return harden({
+    manifest: gameManifest,
+    installations: {
+      game1: restoreRef(game1Ref),
+    },
+  });
+};
